9-strcpy.c

+++ 0x05-pointers_arrays_strings/9-strcpy.c
@@ -8,8 +8,8 @@
  */
 char *_strcpy(char *dest, char *src)
 {
-size_t i;
-size_t n;
+int i;
+int n;
 for (i = 0; i < n; &&src[i] != '\0'; i++)
 dest[i] = src[i];
 for (; i < n; i++)
